+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/packages/access-control/src/index.js
@@ -1,5 +1,0 @@
-/**
- * Exports
- */
-export { default as withAccessControlGroup } from './with-access-control-group';
-