+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/packages/components/src/components/acl-multi-select-control/index.js
@@ -1,3 +1,0 @@
-import '../base-styles/editable-on-focus.scss';
-
-export { default as AccessControlListEditableOnFocusMultiSelectControl } from './acl-editable-on-focus-multi-select-control' ;