+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/packages/components/src/components/schema-mode/index.js
@@ -1,4 +1,0 @@
-export { default as SchemaMode } from './schema-mode' ;
-export { default as SchemaModeControl } from './schema-mode-control' ;
-export { default as SchemaModeControlCard } from './schema-mode-control-card' ;
-export { DEFAULT_SCHEMA_MODE, PUBLIC_SCHEMA_MODE, PRIVATE_SCHEMA_MODE } from './schema-modes';